clean up former evidence implementations
