+++ classes/test_dataset.php
@@ -25,7 +25,4 @@
 
 namespace tool_laaudit;
 class test_dataset extends dataset {
-    protected function collect($data = null) {
-        // TODO: Implement collect() method.
-    }
 }
